+++ coolGuard-access/src/main/java/cn/wnhyang/coolGuard/service/impl/FieldGroupServiceImpl.java
@@ -1,6 +1,5 @@
 package cn.wnhyang.coolGuard.service.impl;
 
-import cn.hutool.core.util.StrUtil;
 import cn.wnhyang.coolGuard.constant.RedisKey;
 import cn.wnhyang.coolGuard.convert.FieldGroupConvert;
 import cn.wnhyang.coolGuard.entity.FieldGroup;
@@ -38,7 +37,9 @@ public class FieldGroupServiceImpl implements FieldGroupService {
     @Transactional(rollbackFor = Exception.class)
     @CacheEvict(value = RedisKey.FIELD_GROUP, allEntries = true)
     public Long createFieldGroup(FieldGroupCreateVO createVO) {
-        validateForCreateOrUpdate(null, createVO.getName());
+        if (fieldGroupMapper.selectByName(createVO.getName()) != null) {
+            throw exception(FIELD_GROUP_NAME_EXIST);
+        }
         FieldGroup fieldGroup = FieldGroupConvert.INSTANCE.convert(createVO);
         fieldGroupMapper.insert(fieldGroup);
         return fieldGroup.getId();
@@ -48,16 +49,34 @@ public class FieldGroupServiceImpl implements FieldGroupService {
     @Transactional(rollbackFor = Exception.class)
     @CacheEvict(value = RedisKey.FIELD_GROUP, allEntries = true)
     public void updateFieldGroup(FieldGroupUpdateVO updateVO) {
-        validateForUpdate(updateVO.getId());
-        FieldGroup fieldGroup = FieldGroupConvert.INSTANCE.convert(updateVO);
-        fieldGroupMapper.updateById(fieldGroup);
+        FieldGroup fieldGroup = fieldGroupMapper.selectById(updateVO.getId());
+        if (fieldGroup == null) {
+            throw exception(FIELD_GROUP_NOT_EXIST);
+        }
+        // 标准，不允许删除
+        if (fieldGroup.getStandard()) {
+            throw exception(FIELD_GROUP_STANDARD);
+        }
+        FieldGroup convert = FieldGroupConvert.INSTANCE.convert(updateVO);
+        fieldGroupMapper.updateById(convert);
     }
 
     @Override
     @Transactional(rollbackFor = Exception.class)
     @CacheEvict(value = RedisKey.FIELD_GROUP, allEntries = true)
     public void deleteFieldGroup(Long id) {
-        validateForDelete(id);
+        FieldGroup fieldGroup = fieldGroupMapper.selectById(id);
+        if (fieldGroup == null) {
+            throw exception(FIELD_GROUP_NOT_EXIST);
+        }
+        // 标准，不允许删除
+        if (fieldGroup.getStandard()) {
+            throw exception(FIELD_GROUP_STANDARD);
+        }
+        Long count = fieldMapper.selectCountByFieldGroupName(fieldGroup.getName());
+        if (count > 0) {
+            throw exception(FIELD_GROUP_HAS_FIELD);
+        }
         fieldGroupMapper.deleteById(id);
     }
 
@@ -77,59 +96,6 @@ public class FieldGroupServiceImpl implements FieldGroupService {
         PageResult<FieldGroupVO> convert = FieldGroupConvert.INSTANCE.convert(pageResult);
         convert.getList().forEach(fieldGroup -> fieldGroup.setCount(fieldMapper.selectCountByFieldGroupName(fieldGroup.getName())));
         return convert;
-    }
-
-    private void validateForDelete(Long id) {
-        validateForUpdate(id);
-        Long count = fieldMapper.selectCountByFieldGroupName(fieldGroupMapper.selectById(id).getName());
-        if (count > 0) {
-            throw exception(FIELD_GROUP_HAS_FIELD);
-        }
-    }
-
-    private void validateForUpdate(Long id) {
-        FieldGroup fieldGroup = fieldGroupMapper.selectById(id);
-        if (fieldGroup == null) {
-            throw exception(FIELD_GROUP_NOT_EXIST);
-        }
-        // 标准，不允许删除
-        if (fieldGroup.getStandard()) {
-            throw exception(FIELD_GROUP_STANDARD);
-        }
-    }
-
-    private void validateForCreateOrUpdate(Long id, String name) {
-        // 校验存在
-        validateExists(id);
-        // 校验名唯一
-        validateNameUnique(id, name);
-    }
-
-    private void validateExists(Long id) {
-        if (id == null) {
-            return;
-        }
-        FieldGroup fieldGroup = fieldGroupMapper.selectById(id);
-        if (fieldGroup == null) {
-            throw exception(FIELD_GROUP_NOT_EXIST);
-        }
-    }
-
-    private void validateNameUnique(Long id, String name) {
-        if (StrUtil.isBlank(name)) {
-            return;
-        }
-        FieldGroup fieldGroup = fieldGroupMapper.selectByName(name);
-        if (fieldGroup == null) {
-            return;
-        }
-        // 如果 id 为空，说明不用比较是否为相同 id 的用户
-        if (id == null) {
-            throw exception(FIELD_GROUP_NAME_EXIST);
-        }
-        if (!fieldGroup.getId().equals(id)) {
-            throw exception(FIELD_GROUP_NAME_EXIST);
-        }
     }
 
 }
